修复积分排行榜显示错误

+++ handler/admin/audit/general.task.edit.go
@@ -21,7 +21,7 @@ func EditGeneralTaskHandler(c *gin.Context) {
 	}
 	if task.CertificateStatus == 1 {
 		var user common.User
-		user.Id = uint64(task.UserId)
+		user.Id = task.UserId
 		_, err := task.CompleteTask(user, task.DeviceId, Service, Config)
 		if CheckErr(err, c) {
 			return
